corrections

+++ InkaArt/InkaArt.Data/Sales/ClientData.cs
@@ -37,7 +37,7 @@ namespace InkaArt.Data.Sales
             return clientUpdateAdapter;
         }
 
-        public int InsertClient(int personType, string name, long ruc, long dni, int priority, int type, int state, string address, int phone, string contact, string email)
+        public int InsertClient(int personType, string name, long ruc, long dni, int priority, int type, int state, string address, long phone, string contact, string email)
         {
             adap = clientAdapter();
             data.Clear();
@@ -167,7 +167,7 @@ namespace InkaArt.Data.Sales
             adap.SelectCommand.Parameters[numParams].NpgsqlValue = id;
         }
 
-        public int UpdateClient(string id, int personType, string name, long ruc, long dni, int priority, int type, int state, string address, int phone, string contact, string email)
+        public int UpdateClient(string id, int personType, string name, long ruc, long dni, int priority, int type, int state, string address, long phone, string contact, string email)
         {
             adap = clientAdapter();
